Feature/feature branch

+++ app/Http/Controllers/Grades/GradeController.php
@@ -3,6 +3,8 @@
 
 namespace App\Http\Controllers\Grades;
 use App\Http\Controllers\Controller;
+use CodeZero\UniqueTranslation\UniqueTranslationRule;
+use App\Models\Classroom;
 
 use App\Http\Requests\StoreGrades;
 use App\Models\Grade;
@@ -31,7 +33,15 @@ class GradeController extends Controller
    */
   public function store(StoreGrades $request)
   {
-      try {
+
+    //   if (Grade::where('Name->ar', $request->Name)->orWhere('Name->en',$request->Name_en)->exists()) {
+
+    //       return redirect()->back()->withErrors(trans('Grades_trans.exists'));
+    //   }
+
+
+ try {
+
           $validated = $request->validated();
           $Grade = new Grade();
           /*
@@ -64,6 +74,7 @@ class GradeController extends Controller
    public function update(StoreGrades $request)
  {
    try {
+
        $validated = $request->validated();
        $Grades = Grade::findOrFail($request->id);
        $Grades->update([
@@ -73,7 +84,6 @@ class GradeController extends Controller
        toastr()->success(trans('messages.Update'));
        return redirect()->route('Grades.index');
    }
-
    catch
    (\Exception $e) {
        return redirect()->back()->withErrors(['error' => $e->getMessage()]);
@@ -89,12 +99,22 @@ class GradeController extends Controller
   public function destroy(Request $request)
   {
 
-    $Grades = Grade::findOrFail($request->id)->delete();
-    toastr()->error(trans('messages.Delete'));
-    return redirect()->route('Grades.index');
+    $MyClass_id = Classroom::where('Grade_id',$request->id)->pluck('Grade_id');
+
+      if($MyClass_id->count() == 0){
+
+          $Grades = Grade::findOrFail($request->id)->delete();
+          toastr()->error(trans('messages.Delete'));
+          return redirect()->route('Grades.index');
+      }
+
+      else{
+
+          toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+          return redirect()->route('Grades.index');
+
+      }
 
   }
 
 }
-
-?>
